# threads-api/swagger/paths.yaml
...
openapi: 3.0.3
info:
  title: Swagger - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:4000
...
paths:
...
  /posts/{post_id}/bookmark:
    post:
      tags:
        - posts
      summary: Bookmark a post
      description: Bookmark a post
      operationId: bookmarkPost
      security:
        - BearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          description: The ID of the bookmark to delete.
          schema:
            type: string
            example: 66d6059974ecbef1214d20ab
      responses:
        default:
          description: Bookmark post successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Bookmark success"
                  data:
                    $ref: '#/components/schemas/Bookmark'
        '422':
          description: Invalid or missing input fields
    delete:
      tags:
        - posts
      summary: Delete a bookmark by ID
      description: Deletes a specific bookmark based on the provided ID.
      security:
        - BearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          description: The ID of the bookmark to delete.
          schema:
            type: string
            example: 66d6059974ecbef1214d20ab
      responses:
        '200':
          description: Successfully deleted the bookmark.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unbookmark successfully"
                  data:
                    $ref: '#/components/schemas/Bookmark'
        '422':
          description: Invalid or missing input fields
  /posts/{post_id}/like:
    post:
      tags:
        - posts
      summary: Like a post
      description: Like a post
      operationId: likePost
      security:
        - BearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          description: ID of the post to remove the like from
          schema:
            type: string
            example: 66d6059974ecbef1214d20ab
      responses:
        default:
          description: Like post successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Like successfully"
                  data:
                    $ref: '#/components/schemas/Like'
        '422':
          description: Invalid or missing input fields
    delete:
      tags:
        - posts
      summary: Remove a like from a post
      description: Deletes the like associated with the specified post ID.
      security:
        - BearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          description: ID of the post to remove the like from
          schema:
            type: string
            example: 66d6059974ecbef1214d20ab
      responses:
        '200':
          description: Like successfully removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unlike successfully"
                  data:
                    $ref: '#/components/schemas/Like'
        '422':
          description: Invalid or missing input fields
...